Move thumbnails related modules into a separate package
